Fix error type associated with Client trait
Instead of having an associated type on the `Client` trait we always use
the `Error` from `subxt`. This makes it possible to use the client trait
generically for both the node client and the memory client without doing
error conversion.

We also changed the error handling behavior of the `MemoryClient`.
Instead of passing errors from the runtime functions we expect them to
be ok and panic otherwise.