Replace SyncClient with ClientWithExecutor
We replace `SyncClient` with `ClientWithExecutor`. Instead of returning
results synchronously like `SyncClient`, `ClientWithExecutor` returns
futures. But these futures are spawned in an executor that is owned by
the client. This means that `ClientWithExecutor` can still be used in
the proxy.

The advantage of the change is that `ClientWithExecutor` unlike
`SyncClient` implements the `Client` trait. This enables us to make the
app proxy generic over the client.