Return call result from 'Client::submit'
We change `TransactionApplied` returned by `Client::submit` so that it
includes the result of the call submitted.

This is accomplished by changing the call argument to `submit` from an
enum to a trait. This new `Call` trait is implemented for all runtime
calls, i.e. the variants that made up the `Call` enum before. The `Call`
trait has an associated `Result` type that specifies the result of a
runtime call.

Currently the result type for all calls is `Result<(), &'static str>`
which maps one to one to the `DispatchResult` of a method in the runtime
definition. In the future we’ll provide proper return values.