ci: Do not pollute /cache in base image
To address an issue first raised in https://github.com/oscoin/ci/pull/41
we change the CI base image so that it does not write data to the
`/cache` directory

To make it easier for other images to use our image as a base image we
also don’t set `RUSTC_WRAPPER` and `CARGO_HOME` in the image. We set
them in the CI script.

We also take the opportunity to update the rust toolchain.