Update state::Project and add core::Project
The state::Project is updated to the latest spec, not containing in
itself the org and id since those will already be stored as the key in
the storage map and can be extracted. The new core::Project type will be
used everywhere else as the status quo of what a 'Project' is in its
entirity.