Withdraw fees for message::UnregisterOrg
I am adding the 'bid' fee to the message itself because, in order to
charge the fee, we need to access the bid (or - at the very least - the
tip) within the registry function handling that message. Placing it in
the TransactionExtra or as an extra field of client::Transaction would
superficially seem better but, in the end, we'd have to change the
'into_runtime_call' to receive a Bid (or, again, a Tip), which would be
no better than add the bid to the message itself and let the whole
pipeline as is.