Implement user registry
Introduces users as first class concept and symmetrical to orgs allows
for registration of them to form a globally unique namespace. It is left
for a later changeset to consolidate the two entities to share the same
id space. See https://github.com/radicle-dev/registry-spec/pull/11 for
the current user semantics according to spec.

Closes #245
Fixes #256