Implement block rewards
We add a block author inherent to the runtime and credit the author with
block rewards and transaction fees.

We don’t implement setting the block author from the CLI yet.

The code enforces some invariants with `panic!`s and `assert!`s. This
follows common substrate practice. (See [`pallet_timestamp`][1] for example.)

[1]: https://github.com/paritytech/substrate/blob/1ea615c0a422fdbc7ecdbc7357bdf7f5bbdc421b/frame/timestamp/src/lib.rs#L149