Test block rewards with emulator
We add tests for the client emulator that assert that block rewards are
properly credited.

To be able to write these tests we need to author blocks in the client
emulator. Before we were just executing transactions.

Now that transactions fees are properly credited by the emulator an
`Endowed` event is dispatched for the block author. This requires us to
adjust the events check in all tests.