cli: Refactor and rename Args as CommandLine
What we now call 'Args' really is the model of our command-line. I was
initially confused with this type, thinking it would capture the
arguments of commands. This change makes it all clear through naming but
also through structure, where it's clear that our command-line is made
of a command and options. Furthermore, have the now 'CommandLine' offer
a 'run' function for improved encapsulation, where main.rs really only
has to connect the dots and know nothing about the internals. At last,
improve other naming and reduce unnecessary verbosity.