core: Make RegistryError derive-try-from-primitive
Doing so allows us to try going from u8 to RegistryError in addition to
the counter part that is already available out of the box. Using this
derivative we don't need to implement it via TryFrom/TryInto, which
saves us time and headspace as that solution would require us to
remember to maintain the mapping u8 -> RegistryError without any
assistance from the compiler.