cli: Improve transaction error propagation
Here we switch to make use of the tx_applied.result Err variant,
wrapping a DispatchError, instead of dropping it from the following
error handling pipeline via CommandError::FailedTransaction.
We now attempt to convert the DispatchError to RegistryError and offer
TransactionError with variants for both (sub)sets, which improves
contextual information and logging.
Note: Changed transaction_applied_ok to be non-generic on the Error type
of the TransactionApplied argument, being explicitly set as
DispatchError, which is the Error type for *all* Message implementations
at the moment and forseeable future.
In the near future, we should change improve error propagation from the
runtime and client so that the CLI doesn't need to handle this to
provide useful information to users, that should be available
beforehand. Issues already exist to tackle that. For now, this solution
solves the issue in the CLI, with special care in mind for our first
real (FFNet) users.