Drop ReturnValue associated type from Message
Big chunck, where the Message trait no longer has associated types
determining the return values of their execution. All messages return
TransactionError and all but CreateCheckpoint return () when successful.
Having that as the only exception, and since we can calculate the value
returned by CreateCheckpoint (CheckpointId), drop that as well.
We are now in a position where we can think of further simplifications.