Make Org members Vec<Id> instead of Vec<AccountId>
We check that authors registering an org are registered users, which was
a missing validation step (#397).

That also means that to authorize actions on orgs, we now need to first
lookup the user associated with the tx author.

When it comes to testing, now most tests need to have the tx author
registering itself as a user.