cli: Check if key-pairs.json exist before migrate
If a user has both the old accounts.json and the new key-pairs.json, the
key-pair storage now assumes that having the older means the latest is
not present. That is fine for all public users, but dangerous for us
developers since we go back and forth with versions of the cli. It is
also useful to have this check for future-compatibility for normal users
anyway, shall one happen to downgrade and move back forth.