Add runtime update policy
We add a policy and describe the process for runtime updates in the
developer handbook.

With this policy the repo now includes to wasm runtime blobs.
`./runtime/latest.wasm` is used by the `dev*` chains.

We also change the execution strategy for all nodes to run both the wasm
and the native code if possible.

See #434.