Improve CI build times with incremental builds
We improve build times on CI by re-enabling incremental builds. This
allows the different builds in the `./ci/run` script to better share
artifacts. We remove the incremental artifacts at the end of the script
because they cannot be used by subsequent builds from a fresh repo
checkout.

We also change the build of docs to use the release profile so that it
can reuse artifacts from the other builds.

Fixes #378