Add control interface for client emulator
Adds `EmulatorControl` and `add_blocks()` for #453. This PR does not yet
at the capability to produce blocks automatically in the background.

A conscious decision was made to expose a separate `EmulatorControl`
struct instead of `Emulator` so that the internal APIs of `Emulator`
(e.g. the `Backend` implementation) are not leaked to the user.