Disable RPC CORS on node configuration
The previous approach of disabling RPC CORS on `RunCmd` did not work
anymore. We need to disable it on the node configuration.

Note that we are only adjusting the config for the full node which is ok
for now. I’d like to use [`Runner::config_mut`][1] but in our substrate
version it has the wrong type.

[1]: https://releases.parity.io/substrate-rustdoc/sc_cli/struct.Runner.html#method.config_mut