Use native runtime for dev node
If we change the runtime code and build the dev node the node should run
the changed native runtime without us having to recompile the genesis
was runtime.

Previously the node would run both the on-chain WASM runtime and the
native one, would detect a difference and use the result from the WASM
runtime. This did not respect the local changes to the runtime.