Static ffnet chain spec
Fixes #493.

Before this change the genesis state was dynamically generated using the
runtime code. Since the runtime code has changed (specifically the
`spec_version`) we were now generating a different genesis state—and
thus genesis block—for the ffnet. As a consequence nodes compiled with
the newer runtime failed to join the ffnet.

With this change we fix the genesis state for the ffnet once and for all
in JSON chain spec and check it into the repo. We also document the
process and remove the unused `ffnet-genesis.wasm` runtime code.