Node binary prints version and commit hash
The version identifier that the node binary prints is now generated from
`git describe`. This allows us to easily identify node versions.

Moreover we adjust the tagging/versioning scheme by replacing dashes in
the date with dots and removing the trailing zero for the first release.
For example `2020.06.12` instead of `2020-06-12-0`. This leads to more
readable `git describe` output. For example `2020.06.09-13-g8ac2839`
instead of `2020-06-09-0-13-g8ac2839`.