Decouple client and substrate runtime modules
The main goal of this change is to decouple the client from the
third-party substrate modules. Specifically we want the client to only
depend on the runtime crate and not to depend on dependencies of the
runtime crate. The motivation behind this is to allow the client to link
against multiple versions of the runtime crate.

A secondary objective that this change achieves is to be more
disciplined in the public API of the runtime crate. We want to reduce
the number of things that are exposed. To this end we move the runtime
construction that declares multiple public types and structs to a
separate module from which we re-export selectively.